Don't uninstall when the add-on is disabled
Uninstall-on-disable is a behavior that made sense for studies that showed up in about:addons, before we could hide study add-ons. Nowadays, there is no user-accessible way outside of a browser console to disable a study. Further, none of the experiment automation tools disable add-ons. At the end of the study, the add-on is uninstalled. Leaving in uninstall-on-disable makes studies more fragile, and serves no useful function anymore.